refactor: move image control
