+++ packages/controls/src/index.ts
@@ -4,6 +4,7 @@ export * from './checkbox'
 export * from './color'
 export * from './combobox'
 export * from './icon-radio-group'
+export * from './image'
 export * from './number'
 export * from './text-input'
 export * from './text-area'
