+++ packages/runtime/src/controls/index.ts
@@ -1,6 +1,5 @@
 export * from './color'
 export * from './control'
-export * from './image'
 export * from './link'
 export * from './list'
 export * from './select'
